Se implementa api gateway y Circuit Breaker en el servicio de ingresos

+++ IncomeService/src/main/java/com/nicolasMorales/IncomeService/models/Product.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 public class Product {
 
-    private int codigo;
+    private Long codigo;
 
     private String name;
