+++ ProductService/src/main/java/com/nicolasMorales/ProductService/repository/IProductRepository.java
@@ -12,6 +12,6 @@ import java.util.UUID;
 public interface IProductRepository  extends JpaRepository<Product, UUID> {
 
     @Query("SELECT p FROM Product p WHERE p.codigo = :codigo")
-    Product findByCodigo(@Param("codigo") int codigo);
+    Product findByCodigo(@Param("codigo") long codigo);
 
 }
